requirements.txt
asgi_redis==0.13.1
asgi_ipc==1.1.0
asgiref==0.13.3
autobahn==0.14.1
channels==0.14.2
daphne==0.13.1
Django==1.9.7
docutils==0.12
msgpack-python==0.4.7
redis==2.10.5
six==1.10.0
statistics==1.0.3.5
Twisted==16.2.0
txaio==2.5.1
zope.interface==4.2.0
django-knocker